create interfaces to smlwriter/parser
